+++ examples/communication_test.cpp
@@ -57,10 +57,10 @@ int main(int argc, char** argv) {
         [&time, &counter, &avg_success_rate, &min_success_rate, &max_success_rate, zero_torques](
             const franka::RobotState& robot_state, franka::Duration period) -> franka::Torques {
           time += period.toMSec();
-          counter++;
-          if (robot_state.control_command_success_rate == 0.0) {
+          if (time == 0.0) {
             return zero_torques;
           }
+          counter++;
 
           if (counter % 100 == 0) {
             std::cout << "#" << counter
@@ -84,7 +84,8 @@ int main(int argc, char** argv) {
 
           // Sending zero torques - if EE is configured correctly, robot should not move
           return zero_torques;
-        });
+        },
+        false, 1000);
   } catch (const franka::Exception& e) {
     std::cout << e.what() << std::endl;
     return -1;