+++ include/franka/rate_limiting.h
@@ -113,6 +113,8 @@ constexpr double kMaxElbowVelocity =
  * @param[in] commanded_values Commanded values of the current time step.
  * @param[in] last_commanded_values Commanded values of the previous time step.
  *
+ * @throw std::invalid_argument if commanded_values are infinite or NaN.
+ *
  * @return Rate-limited vector of desired values.
  */
 std::array<double, 7> limitRate(const std::array<double, 7>& max_derivatives,
@@ -131,6 +133,8 @@ std::array<double, 7> limitRate(const std::array<double, 7>& max_derivatives,
  * @param[in] commanded_velocity Commanded joint velocity of the current time step.
  * @param[in] last_commanded_velocity Commanded joint velocitiy of the previous time step.
  * @param[in] last_commanded_acceleration Commanded joint acceleration of the previous time step.
+ *
+ * @throw std::invalid_argument if commanded_velocity is infinite or NaN.
  *
  * @return Rate-limited desired joint velocity.
  */
@@ -155,6 +159,8 @@ double limitRate(double max_velocity,
  * @param[in] last_commanded_velocity Commanded joint velocity of the previous time step.
  * @param[in] last_commanded_acceleration Commanded joint acceleration of the previous time step.
  *
+ * @throw std::invalid_argument if commanded_position is infinite or NaN.
+ *
  * @return Rate-limited desired joint position.
  */
 double limitRate(double max_velocity,
@@ -178,6 +184,8 @@ double limitRate(double max_velocity,
  * @param[in] last_commanded_velocities Commanded joint velocities of the previous time step.
  * @param[in] last_commanded_accelerations Commanded joint accelerations of the previous time step.
  *
+ * @throw std::invalid_argument if commanded_velocities are infinite or NaN.
+ *
  * @return Rate-limited vector of desired joint velocities.
  */
 std::array<double, 7> limitRate(const std::array<double, 7>& max_velocity,
@@ -200,6 +208,8 @@ std::array<double, 7> limitRate(const std::array<double, 7>& max_velocity,
  * @param[in] last_commanded_positions Commanded joint positions of the current time step.
  * @param[in] last_commanded_velocities Commanded joint positions of the previous time step.
  * @param[in] last_commanded_accelerations Commanded joint velocities of the previous time step.
+ *
+ * @throw std::invalid_argument if commanded_positions are infinite or NaN.
  *
  * @return Rate-limited vector of desired joint positions.
  */
@@ -226,6 +236,8 @@ std::array<double, 7> limitRate(const std::array<double, 7>& max_velocity,
  * @param[in] O_dP_EE_c Commanded end effector twist of the current time step.
  * @param[in] last_O_dP_EE_c Commanded end effector twist of the previous time step.
  * @param[in] last_O_ddP_EE_c Commanded end effector acceleration of the previous time step.
+ *
+ * @throw std::invalid_argument if an element of O_dP_EE_c is infinite or NaN.
  *
  * @return Rate-limited desired end effector twist.
  */
@@ -256,6 +268,8 @@ std::array<double, 6> limitRate(
  * @param[in] last_O_T_EE_c Commanded pose of the previous time step.
  * @param[in] last_O_dP_EE_c Commanded end effector twist of the previous time step.
  * @param[in] last_O_ddP_EE_c Commanded end effector acceleration of the previous time step.
+ *
+ * @throw std::invalid_argument if an element of O_T_EE_c is infinite or NaN.
  *
  * @return Rate-limited desired pose.
  */