+++ CHANGELOG.md
@@ -4,8 +4,25 @@
 
 Requires Panda system version >= 3.0.0
 
- * Added support for using `franka::Model` on Linux ARM and ARM64
+### Added
+
+ * Add support for using `franka::Model` on Linux ARM and ARM64
    (requires Panda system version >= 3.1.0).
+ * Add Cartesian low-pass filter in `lowpass_filter.h`.
+ * Add `control_tools.h` with helper functions for writing control loops.
+
+### Changed
+
+ * **BREAKING** Rename `franka::CartesianPose::hasValidElbow()`
+   to `franka::CartesianPose::hasElbow()`.
+ * **BREAKING** Throw `std::invalid_argument` in `franka::Robot::control` if
+   NaN or infinity values are returned from a control callback.
+ * **BREAKING** Throw `std::invalid_argument` in low-pass filter and rate limiting
+   functions if invalid values are given as function parameters.
+ * **BREAKING** Do not throw exceptions in constructors of control types anymore.
+ * Take arguments by value in `franka::MotionFinished`.
+ * Bug fixes in `communication_test.cpp`.
+ * Export requirements for C++ features as CMake target compile features.
 
 ## 0.6.0 - 2019-02-06